Move locks for largest socket fd
The call to getLargestSocket places locks on the socket resource, but
then loops thorugh the sockets later when calling select. Since the
function that calls select is the only caller for getLargestSocket, it
makes sense to lock the resource bundled together in case the largest
socket is somehow manipulated or removed before `select` is called.