NATHP Server should own handlers
The AcceptHandler and ProcessHandler passed to the nathp::server
contstructor should be passed by rvalue so that the object is actually
passed to and owned by the server. After passing these objects off to
the server, there's no sense in verifying their use outside of the
server object and thus become temporaries and should remain unused.

Future implentations that seek to dynamically change the functions the
server calls should be an super implenetation